Alignment wip
Still doesn't work when alignment > 16 bytes.

+++ src/main/java/com/quaxt/mcc/asm/ByteArray.java
@@ -1,6 +1,6 @@
 package com.quaxt.mcc.asm;
 
-public record ByteArray(long size, long alignment) implements TypeAsm {
+public record ByteArray(long size, int alignment) implements TypeAsm {
     public String suffix() {
         return "q";
     }
